requirements.txt
livekit-agents[silero,openai,cartesia]
python-dotenv
faster-whisper
google-cloud-aiplatform
livekit-plugins-google
